+++ test/diffclj/core_test.clj
@@ -228,5 +228,12 @@
       (is
        (=
         (simplify (deriv '(atan x)))
-        '(/ 1 (+ 1 (pow x 2))))))))
+        '(/ 1 (+ 1 (pow x 2))))))
+
+
+    (binding [*ns*          (find-ns 'diffclj.core)]
+      (is
+       (=
+        (simplify (deriv '(acot x)))
+        '(/ -1 (+ 1 (pow x 2))))))))